+++ test-utils/test-utils-core/src/main/java/uk/gov/justice/services/test/utils/core/matchers/JsonEnvelopeListMatcher.java
@@ -16,7 +16,22 @@ import org.hamcrest.collection.IsIterableContainingInOrder;
  * Matches a List of JsonEnvelopes.  This is very similar to {@link JsonEnvelopeStreamMatcher} and
  * is used internally by the JsonEnvelopeStreamMatcher.
  *
- * See {@link JsonEnvelopeStreamMatcher} for example usage
+ * <pre>
+ *  {@code
+ *      assertThat(eventList, listContaining(
+ *              jsonEnvelope(
+ *                      metadata()
+ *                          .withCausationIds(commandId)
+ *                          .withName(EVENT_NAME),
+ *                      payloadIsJson(allOf(
+ *                          withJsonPath("$.recipeId", equalTo(RECIPE_ID.toString())),
+ *                          withJsonPath("$.name", equalTo(RECIPE_NAME)),
+ *                          withJsonPath("$.glutenFree", equalTo(GULTEN_FREE)),
+ *                          withJsonPath("$.ingredients", empty())
+ *                      ))))
+ *      );
+ * }
+ * </pre>
  */
 public class JsonEnvelopeListMatcher extends TypeSafeDiagnosingMatcher<List<JsonEnvelope>> {
 
@@ -32,7 +47,7 @@ public class JsonEnvelopeListMatcher extends TypeSafeDiagnosingMatcher<List<Json
 
     @Override
     public void describeTo(final Description description) {
-        description.appendText("JsonEnvelope Stream that contains: ");
+        description.appendText("JsonEnvelope List that contains: ");
         stream(matchers).forEach(description::appendDescriptionOf);
     }