+++ test-utils/test-utils-core/src/test/java/uk/gov/justice/services/test/utils/core/matchers/JsonEnvelopeListMatcherTest.java
@@ -11,7 +11,7 @@ import static uk.gov.justice.services.messaging.DefaultJsonEnvelope.envelope;
 import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataWithRandomUUID;
 import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopeMatcher.jsonEnvelope;
 import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopeMetadataMatcher.metadata;
-import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopePayloadMatcher.payLoad;
+import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopePayloadMatcher.payloadIsJson;
 
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
@@ -37,13 +37,13 @@ public class JsonEnvelopeListMatcherTest {
         assertThat(events, JsonEnvelopeListMatcher.listContaining(
                 jsonEnvelope(
                         metadata().withName("event.action"),
-                        payLoad().isJson(allOf(
+                        payloadIsJson(allOf(
                                 withJsonPath("$.someId", equalTo(ID_1.toString())),
                                 withJsonPath("$.name", equalTo(NAME_1)))
                         )),
                 jsonEnvelope(
                         metadata().withName("event.action"),
-                        payLoad().isJson(allOf(
+                        payloadIsJson(allOf(
                                 withJsonPath("$.someId", equalTo(ID_2.toString())),
                                 withJsonPath("$.name", equalTo(NAME_2)))
                         ))
@@ -60,7 +60,7 @@ public class JsonEnvelopeListMatcherTest {
         assertThat(events, JsonEnvelopeListMatcher.listContaining(
                 jsonEnvelope(
                         metadata().withName("event.action"),
-                        payLoad().isJson(allOf(
+                        payloadIsJson(allOf(
                                 withJsonPath("$.someId", equalTo(ID_1.toString())),
                                 withJsonPath("$.name", equalTo(NAME_1)))
                         ))
@@ -76,13 +76,13 @@ public class JsonEnvelopeListMatcherTest {
         assertThat(events, JsonEnvelopeListMatcher.listContaining(
                 jsonEnvelope(
                         metadata().withName("event.action"),
-                        payLoad().isJson(allOf(
+                        payloadIsJson(allOf(
                                 withJsonPath("$.someId", equalTo(ID_1.toString())),
                                 withJsonPath("$.name", equalTo(NAME_1)))
                         )),
                 jsonEnvelope(
                         metadata().withName("event.action"),
-                        payLoad().isJson(allOf(
+                        payloadIsJson(allOf(
                                 withJsonPath("$.someId", equalTo(ID_2.toString())),
                                 withJsonPath("$.name", equalTo(NAME_2)))
                         ))