+++ test-utils/test-utils-core/src/test/java/uk/gov/justice/services/test/utils/core/matchers/JsonEnvelopeMatcherTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertThat;
 import static uk.gov.justice.services.messaging.DefaultJsonEnvelope.envelope;
 import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataWithRandomUUID;
 import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopeMetadataMatcher.metadata;
-import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopePayloadMatcher.payLoad;
+import static uk.gov.justice.services.test.utils.core.matchers.JsonEnvelopePayloadMatcher.payloadIsJson;
 
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
@@ -25,7 +25,7 @@ public class JsonEnvelopeMatcherTest {
     public void shouldMatchJsonEnvelope() throws Exception {
         assertThat(jsonEnvelope(), JsonEnvelopeMatcher.jsonEnvelope(
                 metadata().withName("event.action"),
-                payLoad().isJson(allOf(
+                payloadIsJson(allOf(
                         withJsonPath("$.someId", equalTo(ID.toString())),
                         withJsonPath("$.name", equalTo(NAME)))
                 )));
@@ -35,7 +35,7 @@ public class JsonEnvelopeMatcherTest {
     public void shouldFailToMatchDifferentMetadata() throws Exception {
         assertThat(jsonEnvelope(), JsonEnvelopeMatcher.jsonEnvelope(
                 metadata().withName("event.not.match"),
-                payLoad().isJson(allOf(
+                payloadIsJson(allOf(
                         withJsonPath("$.someId", equalTo(ID.toString())),
                         withJsonPath("$.name", equalTo(NAME)))
                 )));
@@ -45,7 +45,7 @@ public class JsonEnvelopeMatcherTest {
     public void shouldFailToMatchDifferentPayload() throws Exception {
         assertThat(jsonEnvelope(), JsonEnvelopeMatcher.jsonEnvelope(
                 metadata().withName("event.action"),
-                payLoad().isJson(allOf(
+                payloadIsJson(allOf(
                         withJsonPath("$.someId", equalTo(randomUUID().toString())),
                         withJsonPath("$.name", equalTo(NAME)))
                 )));